requirements.txt
streamlit>=1.32.0
mysql-connector-python>=9.0.0
pandas>=2.0.0
